phase_3/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
uvicorn==0.23.0
sqlalchemy
python-dotenv
psycopg2
sqlalchemy-cockroachdb
requests
amqpstorm
pika
celery[amqp]